+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.18.0] - (16.10.2024)
+
+## Changes
+
+- added frost protect function for MiElHVAC
+- config.schema update
+- redme update
+- cleanup
+
 ## [0.17.0] - (14.10.2024)
 
 ## Changes